Allow opening multiple files with :e
